requirements.txt
# Core Data & Computation
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
requests>=2.31.0

# Technical Analysis
ta>=0.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.14.0
joblib>=1.3.0

# Visualization & Dashboard
plotly==5.19.0

# Scheduling
schedule==1.2.1

# Utilities
python-dateutil==2.8.2
fastapi
uvicorn

# Multi-Agent Upgrade
google-generativeai>=0.3.0
websockets>=12.0
pydantic>=2.0
